split things out a bit and add a typecheck script
