+++ src/utils/constants.ts
@@ -1,0 +1,10 @@
+export const PLAYER_ONE = "red";
+export const PLAYER_TWO = "yellow";
+export const PLAYER_THREE = "purple";
+export const PLAYER_FOUR = "green";
+export const allPlayers = [
+  PLAYER_ONE,
+  PLAYER_TWO,
+  PLAYER_THREE,
+  PLAYER_FOUR,
+] as const;
